backend/requirements.txt
fastapi==0.115.9
uvicorn==0.40.0
faiss-cpu==1.13.2
PyMuPDF==1.27.1
python-docx==1.2.0
huggingface_hub==0.27.1
numpy==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.20